requirements-dev.txt
django-debug-toolbar==6.2.0
django-extensions==4.1